refactor app and api
