+++ api/requirements.txt
@@ -41,6 +41,7 @@ regex==2024.9.11
 requests==2.32.3
 sniffio==1.3.1
 SQLAlchemy==2.0.34
+sse-starlette==2.1.3
 starlette==0.38.5
 tenacity==8.5.0
 tiktoken==0.7.0
